requirements.txt
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.7
django-autoslug==1.9.3
django-push==1.0
django-rest-framework==0.1.0
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
